Thesis-Project/requirements.txt
opencv-python>=4.9.0.80
torch>=2.2.1 # remove this when installing
ipython>=8.22.2
pillow>=10.2.0
scikit-learn>=1.4.1.post1
torchvision>=0.17.1 # remove this when installing
isort>=5.13.2
matplotlib>=3.8.3
jsonschema>=4.21.1
wandb>=0.16.6
numpy>=1.26.4
transformers==4.41.0
huggingface-hub==0.23.0